+++ src/main/scala/com/raphtory/core/components/graphbuilder/BuilderExecutor.scala
@@ -4,12 +4,13 @@ import akka.cluster.pubsub.DistributedPubSub
 import akka.cluster.pubsub.DistributedPubSubMediator
 import akka.util.Timeout
 import com.raphtory.core.components.RaphtoryActor
-import com.raphtory.core.components.graphbuilder.BuilderExecutor.CommonMessage.{DataFinishedSync, KeepAlive, BuilderTimeSync, StartUp, TimeBroadcast}
+import com.raphtory.core.components.graphbuilder.BuilderExecutor.CommonMessage.{BuilderTimeSync, DataFinishedSync, KeepAlive, StartUp, TimeBroadcast}
 import com.raphtory.core.components.graphbuilder.BuilderExecutor.State
 import com.raphtory.core.components.spout.SpoutAgent.CommonMessage.{AllocateTuple, DataFinished, NoWork, SpoutOnline, WorkPlease}
-import com.raphtory.core.model.communication._
+import com.raphtory.core.implementations.objectgraph.messaging._
 import akka.pattern.ask
-import com.raphtory.core.components.orchestration.raphtoryleader.WatchDog.Message.{ClusterStatusRequest, ClusterStatusResponse, BuilderUp}
+import com.raphtory.core.components.orchestration.raphtoryleader.WatchDog.Message.{BuilderUp, ClusterStatusRequest, ClusterStatusResponse}
+import com.raphtory.core.model.graph.{GraphUpdate, TrackedGraphUpdate}
 
 import scala.collection.mutable
 import scala.collection.parallel.mutable.ParTrieMap
