+++ src/main/scala/com/raphtory/core/implementations/objectgraph/entities/external/ObjectEdge.scala
@@ -1,9 +1,9 @@
-package com.raphtory.core.model.implementations.objectgraph.entities.external
+package com.raphtory.core.implementations.objectgraph.entities.external
 
-import com.raphtory.core.model.communication.VertexMessage
+import com.raphtory.core.implementations.objectgraph.ObjectGraphLens
+import com.raphtory.core.implementations.objectgraph.entities.internal.RaphtoryEdge
+import com.raphtory.core.model.graph.VertexMessage
 import com.raphtory.core.model.graph.visitor.Edge
-import com.raphtory.core.model.implementations.objectgraph.ObjectGraphLens
-import com.raphtory.core.model.implementations.objectgraph.entities.internal.RaphtoryEdge
 
 class ObjectEdge(edge: RaphtoryEdge, id: Long, view: ObjectGraphLens) extends ObjectEntity(edge,view) with Edge {
   def ID() = id
