+++ src/main/scala/com/raphtory/core/implementations/objectgraph/entities/external/ObjectEntity.scala
@@ -1,8 +1,8 @@
-package com.raphtory.core.model.implementations.objectgraph.entities.external
+package com.raphtory.core.implementations.objectgraph.entities.external
 
+import com.raphtory.core.implementations.objectgraph.ObjectGraphLens
+import com.raphtory.core.implementations.objectgraph.entities.internal.RaphtoryEntity
 import com.raphtory.core.model.graph.visitor.{EntityVisitor, HistoricEvent}
-import com.raphtory.core.model.implementations.objectgraph.ObjectGraphLens
-import com.raphtory.core.model.implementations.objectgraph.entities.internal.RaphtoryEntity
 
 import scala.collection.mutable
 import scala.collection.parallel.mutable.ParTrieMap
