+++ src/main/scala/com/raphtory/core/implementations/objectgraph/entities/external/ObjectVertex.scala
@@ -1,9 +1,10 @@
-package com.raphtory.core.model.implementations.objectgraph.entities.external
+package com.raphtory.core.implementations.objectgraph.entities.external
 
-import com.raphtory.core.model.communication.{VertexMessage, VertexMultiQueue}
+import com.raphtory.core.implementations.objectgraph.ObjectGraphLens
+import com.raphtory.core.implementations.objectgraph.entities.internal.RaphtoryVertex
+import com.raphtory.core.implementations.objectgraph.messaging.VertexMultiQueue
+import com.raphtory.core.model.graph.VertexMessage
 import com.raphtory.core.model.graph.visitor.{Edge, EntityVisitor, Vertex}
-import com.raphtory.core.model.implementations.objectgraph.ObjectGraphLens
-import com.raphtory.core.model.implementations.objectgraph.entities.internal.RaphtoryVertex
 
 import scala.collection.parallel.mutable.ParTrieMap
 import scala.reflect.ClassTag
