+++ src/main/scala/com/raphtory/core/implementations/objectgraph/entities/internal/RaphtoryEntity.scala
@@ -1,4 +1,4 @@
-package com.raphtory.core.model.implementations.objectgraph.entities.internal
+package com.raphtory.core.implementations.objectgraph.entities.internal
 
 import scala.collection.mutable
 import scala.collection.parallel.mutable.ParTrieMap
