+++ src/main/scala/com/raphtory/core/implementations/objectgraph/messaging/VertexMessageHandler.scala
@@ -1,9 +1,10 @@
-package com.raphtory.core.model.communication
+package com.raphtory.core.implementations.objectgraph.messaging
 
 import java.util.concurrent.atomic.AtomicInteger
 import akka.actor.ActorRef
 import akka.cluster.pubsub.DistributedPubSubMediator
 import com.raphtory.core.components.RaphtoryActor.totalPartitions
+import com.raphtory.core.model.graph.VertexMessage
 
 import scala.collection.mutable
