+++ src/main/scala/com/raphtory/core/model/graph/GraphPartition.scala
@@ -1,9 +1,9 @@
 package com.raphtory.core.model.graph
 
 import com.raphtory.core.components.RaphtoryActor.totalPartitions
-import com.raphtory.core.model.communication._
+import com.raphtory.core.implementations.objectgraph.entities.internal.RaphtoryVertex
+import com.raphtory.core.implementations.objectgraph.messaging._
 import com.raphtory.core.model.graph.visitor.Vertex
-import com.raphtory.core.model.implementations.objectgraph.entities.internal.RaphtoryVertex
 import com.typesafe.scalalogging.LazyLogging
 
 import scala.collection.concurrent.TrieMap
