+++ src/main/scala/com/raphtory/core/model/graph/GraphUpdates.scala
@@ -1,8 +1,4 @@
-package com.raphtory.core.model.communication
-
-import com.raphtory.core.model.implementations.objectgraph.entities.internal.RaphtoryEdge
-
-import scala.collection.mutable
+package com.raphtory.core.model.graph
 
 /**
   * Created by Mirate on 30/05/2017.
