+++ src/main/scala/com/raphtory/core/model/graph/visitor/Edge.scala
@@ -1,8 +1,8 @@
 package com.raphtory.core.model.graph.visitor
 
-import com.raphtory.core.model.communication.VertexMessage
-import com.raphtory.core.model.implementations.objectgraph.ObjectGraphLens
-import com.raphtory.core.model.implementations.objectgraph.entities.internal.RaphtoryEdge
+import com.raphtory.core.implementations.objectgraph.ObjectGraphLens
+import com.raphtory.core.implementations.objectgraph.entities.internal.RaphtoryEdge
+import com.raphtory.core.model.graph.VertexMessage
 
 trait Edge extends EntityVisitor {
