+++ src/main/scala/com/raphtory/core/model/graph/visitor/EntityVisitor.scala
@@ -1,7 +1,7 @@
 package com.raphtory.core.model.graph.visitor
 
-import com.raphtory.core.model.implementations.objectgraph.ObjectGraphLens
-import com.raphtory.core.model.implementations.objectgraph.entities.internal.RaphtoryEntity
+import com.raphtory.core.implementations.objectgraph.entities.internal.RaphtoryEntity
+
 
 import scala.collection.mutable
 import scala.collection.parallel.mutable.ParTrieMap
