+++ src/main/scala/com/raphtory/dev/blockchain/graphbuilders/chab_C2C_GB.scala
@@ -1,7 +1,8 @@
 package com.raphtory.dev.blockchain.graphbuilders
 
 import com.raphtory.core.components.graphbuilder.GraphBuilder
-import com.raphtory.core.model.communication.{Type, _}
+import com.raphtory.core.implementations.objectgraph.messaging._
+import com.raphtory.core.model.graph.{FloatProperty, LongProperty, Properties, StringProperty, Type}
 
 class chab_C2C_GB extends GraphBuilder[String]{
