+++ src/main/scala/com/raphtory/dev/gab/graphbuilders/GabUserGraphBuilder.scala
@@ -1,9 +1,9 @@
 package com.raphtory.dev.gab.graphbuilders
 
 import java.text.SimpleDateFormat
-
 import com.raphtory.core.components.graphbuilder.GraphBuilder
-import com.raphtory.core.model.communication.{Type, _}
+import com.raphtory.core.implementations.objectgraph.messaging._
+import com.raphtory.core.model.graph.Type
 
 // The lines sent by the Gab mining spout are read and processed accordingly.
 //In this builder we needed to transform the data that was sent by the spout by turning it into a epoch value (long value)
