+++ src/main/scala/com/raphtory/dev/lotr/LOTRGraphBuilder.scala
@@ -1,7 +1,8 @@
 package com.raphtory.dev.lotr
 
 import com.raphtory.core.components.graphbuilder.GraphBuilder
-import com.raphtory.core.model.communication._
+import com.raphtory.core.implementations.objectgraph.messaging._
+import com.raphtory.core.model.graph.{ImmutableProperty, Properties, Type}
 
 class LOTRGraphBuilder extends GraphBuilder[String]{
