+++ src/main/scala/com/raphtory/dev/wordSemantic/graphbuilders/CoMatGBFiltered.scala
@@ -1,7 +1,8 @@
 package com.raphtory.dev.wordSemantic.graphbuilders
 
 import com.raphtory.core.components.graphbuilder.GraphBuilder
-import com.raphtory.core.model.communication._
+import com.raphtory.core.implementations.objectgraph.messaging._
+import com.raphtory.core.model.graph.{FloatProperty, Properties, StringProperty}
 
 import scala.collection.mutable
